+++ Plugin/package/CHANGELOG.md
@@ -1,3 +1,8 @@
+# v1.0.6
+- Added damage configuration
+- Reduced default damage from 20 to 15
+- Made spawn configuration more flexible through own helper library
+
 # v1.0.5
 - Updated for v80
 - Improved spawn configuration handling